# [SPARK-26711][SQL] Lazily convert string values to BigDecimal during JSON schema inference